requirements.txt
aiogram==2.20
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
